JRobin 1.4.0 - Fixed bug #0000042 - Minor tweaks

+++ src/org/jrobin/graph/Stack.java
@@ -63,7 +63,7 @@ class Stack extends PlotDef
 	 * @param stackValues Datapoint values of previous PlotDefs, used to stack on if necessary.
 	 * @param lastPlotType Type of the previous PlotDef, used to determine PlotDef type of a stack.
 	 */
-	void draw( ChartGraphics g, int[] xValues, int[] stackValues, int lastPlotType ) throws RrdException
+	void draw( ChartGraphics g, int[] xValues, double[] stackValues, int lastPlotType ) throws RrdException
 	{
 		PlotDef stack = null;
